requirements.txt
python-dotenv==1.0.1
typer==0.12.5
tqdm==4.66.5
sacrebleu==2.4.3
rouge-score==0.1.2
pandas==2.2.2
google-generativeai
pyyaml